requirements.txt
Django==3.2.16
djangorestframework==3.13.1
mysqlclient==2.1.1
celery==5.2.7
redis==4.3.4
django-admin-sortable2==1.0.4
gunicorn==20.1.0
drf-yasg==1.21.4
python-dotenv==0.21.0
django-model-utils==4.2.0
Faker==19.3.1
